添加账户和支出的关联性

+++ app/src/main/java/com/lizl/wtmg/db/model/PropertyModel.kt
@@ -22,4 +22,4 @@ data class PropertyModel(
         var showInTotal: Boolean,
 
         @ColumnInfo
-        var amount: Int)
+        var amount: Float)
